Move select demo code to a separate file
